+++ backend/internal/handlers/review/service.go
@@ -2,47 +2,11 @@ package review
 
 import (
 	"context"
-	"time"
 
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
 )
-
-type ReviewDocument struct {
-	ID        primitive.ObjectID       `bson:"_id,omitempty" json:"_id,omitempty"`
-	Rating    Rating                   `bson:"rating" json:"rating"`
-	Picture   string                   `bson:"picture" json:"picture"`
-	Content   string                   `bson:"content" json:"content"`
-	Reviewer  Reviewer                 `bson:"reviewer" json:"reviewer"`
-	Timestamp time.Time                `bson:"timestamp" json:"timestamp"`
-	Comments  []map[string]interface{} `bson:"comments,omitempty" json:"comments,omitempty"`
-	MenuItem  string                   `bson:"menuItem" json:"menuItem"`
-}
-
-// Rating is a nested struct in ReviewDocument.
-type Rating struct {
-	Portion int  `bson:"portion" json:"portion"`
-	Taste   int  `bson:"taste" json:"taste"`
-	Value   int  `bson:"value" json:"value"`
-	Overall int  `bson:"overall" json:"overall"`
-	Return  bool `bson:"return" json:"return"`
-}
-
-// Reviewer is a nested struct in ReviewDocument.
-type Reviewer struct {
-	ID       string `bson:"id"       json:"id"`
-	PFP      string `bson:"pfp"      json:"pfp"`
-	Username string `bson:"username" json:"username"`
-}
-
-/*
-Review Service to be used by Review Handler to interact with the
-Database layer of the application
-*/
-type Service struct {
-	reviews *mongo.Collection
-}
 
 // newService receives the map of collections and picks out reviews
 func newService(collections map[string]*mongo.Collection) *Service {
@@ -178,4 +142,12 @@ func (s *Service) DeleteReview(id primitive.ObjectID) error {
 
 	_, err := s.reviews.DeleteOne(ctx, filter)
 	return err
+}
+
+func (s *Service) CreateComment( comment CommentDocument) error {
+	ctx := context.Background()
+	filter := bson.M{"_id": comment.Review}
+	update := bson.M{"$push": bson.M{"comments": comment}}
+	s.reviews.FindOneAndUpdate(ctx, filter, update)
+	return nil
 }
